Captura de datos y vista en consola. modificacion en los tipos de datos en operaciones

+++ OperacionesBinarias.java
@@ -1,5 +1,5 @@
 public class OperacionesBinarias {
-    public static int multiplicacion(int a, int b){
+    public static double multiplicacion(double a, double b){
         return (a*b);
     }
     public static double division(double c, double d){
